chore: 테스트 클래스 이름변경
